// backend/src/migrateTables/types.ts
type CaseRawType = {
  internal_case_id?: string;
  gender?: string;
  race?: string;
  ethnicity?: string;
  age_at_crime?: string;
  age_at_arrest?: string;
  court_name?: string;
  court_ori?: string;
  county_name?: string;
  district?: string;
  region?: string;
  court_type?: string;
  judge_name?: string;
  offense_month?: string;
  offense_year?: string;
  arrest_month?: string;
  arrest_year?: string;
  arrest_type?: string;
  top_arrest_law?: string;
  top_arrest_article_section?: string;
  top_arrest_attempt_indicator?: string;
  top_charge_at_arrest?: string;
  top_charge_severity_at_arrest?: string;
  top_charge_weight_at_arrest?: string;
  top_charge_at_arrest_violent_felony_ind?: string;
  case_type?: string;
  first_arraign_date?: string;
  top_arraign_law?: string;
  top_arraign_article_section?: string;
  top_arraign_attempt_indicator?: string;
  top_charge_at_arraign?: string;
  top_severity_at_arraign?: string;
  top_charge_weight_at_arraign?: string;
  top_charge_at_arraign_violent_felony_ind?: string;
  hate_crime_ind?: string;
  arraign_charge_category?: string;
  representation_type?: string;
  app_count_arraign_to_dispo_released?: string;
  app_count_arraign_to_dispo_detained?: string;
  app_count_arraign_to_dispo_total?: string;
  def_attended_sched_pretrials?: string;
  remanded_to_jail_at_arraign?: string;
  ror_at_arraign?: string;
  bail_set_and_posted_at_arraign?: string;
  bail_set_and_not_posted_at_arraign?: string;
  nmr_at_arraign?: string;
  release_decision_at_arraign?: string;
  representation_at_securing_order?: string;
  pretrial_supervision_at_arraign?: string;
  contact_pretrial_service_agency?: string;
  electronic_monitoring?: string;
  travel_restrictions?: string;
  passport_surrender?: string;
  no_firearms_or_weapons?: string;
  maintain_employment?: string;
  maintain_housing?: string;
  maintain_school?: string;
  placement_in_mandatory_program?: string;
  removal_to_hospital?: string;
  obey_order_of_protection?: string;
  obey_court_conditions_family_offense?: string;
  other_nmr?: string;
  order_of_protection?: string;
  first_bail_set_cash?: string;
  first_bail_set_credit?: string;
  first_insurance_company_bail_bond?: string;
  first_secured_surety_bond?: string;
  first_secured_app_bond?: string;
  first_unsecured_surety_bond?: string;
  first_unsecured_app_bond?: string;
  first_partially_secured_surety_bond?: string;
  partially_secured_surety_bond_perc?: string;
  first_partially_secured_app_bond?: string;
  partially_secured_app_bond_perc?: string;
  bail_made_indicator?: string;
  warrant_ordered_btw_arraign_and_dispo?: string;
  dat_wo_ws_prior_to_arraign?: string;
  first_bench_warrant_date?: string;
  non_stayed_wo?: string;
  num_of_stayed_wo?: string;
  num_of_row?: number;
  docket_status?: string;
  disposition_type?: string;
  disposition_detail?: string;
  dismissal_reason?: string;
  disposition_date?: string;
  most_severe_sentence?: string;
  top_conviction_law?: string;
  top_conviction_article_section?: string;
  top_conviction_attempt_indicator?: string;
  top_charge_at_conviction?: string;
  top_charge_severity_at_conviction?: string;
  top_charge_weight_at_conviction?: string;
  top_charge_at_conviction_violent_felony_ind?: string;
  days_arraign_remand_first_released?: string;
  known_days_in_custody?: string;
  days_arraign_bail_set_to_first_posted?: string;
  days_arraign_bail_set_to_first_release?: string;
  days_arraign_to_dispo?: string;
  ucmslivedate?: string;
  prior_vfo_cnt?: string;
  prior_nonvfo_cnt?: string;
  prior_misd_cnt?: string;
  pend_vfo?: string;
  pend_nonvfo?: string;
  pend_misd?: string;
  supervision?: string;
  rearrest?: string;
  rearrest_date?: string;
  rearrest_firearm?: string;
  rearrest_date_firearm?: string;
  arr_cycle_id?: string;
  case_raw_id?: string;
};